+++ examples/vike-app/.test-dev.test.ts
@@ -2,5 +2,5 @@ import { testRun } from "./.testRun";
 
 process.env.VITE_CONFIG = "{server:{port:3000}}";
 testRun("pnpm run dev", {
-  serverIsReadyMessage: "Local:",
+  serverIsReadyMessage: "Server running",
 });
