+++ examples/vike-app/package.json
@@ -16,6 +16,7 @@
   "devDependencies": {
     "@auth/core": "^0.38.0",
     "@types/node": "^18.19.70",
+    "@universal-middleware/core": "^0.4.4",
     "typescript": "^5.8.2",
     "vike-cloudflare": "^0.1.7",
     "wrangler": "^3.112.0"
