+++ packages/vike-cloudflare/assets/hono.js
@@ -1,6 +1,6 @@
 import "virtual:@brillout/vite-plugin-server-entry:serverEntry";
-import { Hono } from "hono";
-import app from "virtual:vike-cloudflare-entry";
+import { Hono } from "vike-cloudflare/hono";
+import app from "virtual:vike-cloudflare:user-entry";
 
 const worker = new Hono();
