+++ packages/vike-cloudflare/package.json
@@ -16,13 +16,13 @@
   ],
   "author": "Joël Charles <joel.charles91@gmail.com>",
   "dependencies": {
+    "hono": "^4.7.4",
     "vike-server": "0.3.7-commit-fe2502b"
   },
   "devDependencies": {
     "@brillout/release-me": "^0.4.3",
     "@hattip/adapter-cloudflare-workers": "^0.0.49",
     "esbuild-plugin-raw": "^0.1.8",
-    "hono": "^4.7.4",
     "tsup": "^8.4.0",
     "typescript": "^5.8.2",
     "vike": "0.4.224-commit-f0d0f8a",
