+++ packages/vike-cloudflare/src/config.ts
@@ -1,6 +1,7 @@
 export { config as default };
 
-import { pages as plugin, type VikeCloudflarePagesOptions } from "./index";
+import { pages as plugin } from "./plugins";
+import type { VikeCloudflarePagesOptions } from "./types";
 import type { Config } from "vike/types";
 import server from "vike-server/plugin";
 
@@ -13,7 +14,7 @@ const config = {
     plugins: [
       ...plugin(),
       server({
-        entry: "virtual:vike-cloudflare-auto",
+        entry: "virtual:vike-cloudflare:auto-entry",
         // We're using rollup's noExternal instead
         standalone: false,
         runtime: "workerd",
