+++ packages/vike-cloudflare/src/types.ts
@@ -1,0 +1,7 @@
+export type SupportedServers = "hono" | "hono-dev" | "hattip";
+
+export interface VikeCloudflarePagesOptions {
+  server?: {
+    entry: string;
+  };
+}
